Rename to entity
