+++ src/main/kotlin/no/hvl/dat251/fjelltur/entity/Account.kt
@@ -1,4 +1,4 @@
-package no.hvl.dat251.fjelltur.model
+package no.hvl.dat251.fjelltur.entity
 
 import no.hvl.dat251.fjelltur.dto.AccountId
 import org.hibernate.annotations.GenericGenerator
