+++ src/main/kotlin/no/hvl/dat251/fjelltur/entity/GPSLocation.kt
@@ -1,4 +1,4 @@
-package no.hvl.dat251.fjelltur.model
+package no.hvl.dat251.fjelltur.entity
 
 import java.time.OffsetDateTime
 import javax.persistence.Column
