+++ src/main/kotlin/no/hvl/dat251/fjelltur/entity/Rule.kt
@@ -1,4 +1,4 @@
-package no.hvl.dat251.fjelltur.model
+package no.hvl.dat251.fjelltur.entity
 
 import no.hvl.dat251.fjelltur.dto.RuleId
 import org.hibernate.annotations.GenericGenerator
