+++ src/main/kotlin/no/hvl/dat251/fjelltur/repository/RuleRepository.kt
@@ -1,6 +1,6 @@
 package no.hvl.dat251.fjelltur.repository
 
-import no.hvl.dat251.fjelltur.model.Rule
+import no.hvl.dat251.fjelltur.entity.Rule
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface RuleRepository : JpaRepository<Rule, String> {
