+++ src/main/kotlin/no/hvl/dat251/fjelltur/repository/TripRepository.kt
@@ -1,7 +1,7 @@
 package no.hvl.dat251.fjelltur.repository
 
-import no.hvl.dat251.fjelltur.model.Account
-import no.hvl.dat251.fjelltur.model.Trip
+import no.hvl.dat251.fjelltur.entity.Account
+import no.hvl.dat251.fjelltur.entity.Trip
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
