+++ src/main/kotlin/no/hvl/dat251/fjelltur/service/RuleService.kt
@@ -4,9 +4,9 @@ import no.hvl.dat251.fjelltur.ADMIN_ROLE
 import no.hvl.dat251.fjelltur.CREATE_RULE_PERMISSION
 import no.hvl.dat251.fjelltur.dto.CreateDistanceRuleRequest
 import no.hvl.dat251.fjelltur.dto.CreateTimeRuleRequest
-import no.hvl.dat251.fjelltur.model.DistanceRule
-import no.hvl.dat251.fjelltur.model.Rule
-import no.hvl.dat251.fjelltur.model.TimeRule
+import no.hvl.dat251.fjelltur.entity.DistanceRule
+import no.hvl.dat251.fjelltur.entity.Rule
+import no.hvl.dat251.fjelltur.entity.TimeRule
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.security.access.prepost.PreAuthorize
