+++ src/main/kotlin/no/hvl/dat251/fjelltur/service/impl/RuleServiceImpl.kt
@@ -3,9 +3,9 @@ package no.hvl.dat251.fjelltur.service.impl
 import no.hvl.dat251.fjelltur.dto.CreateDistanceRuleRequest
 import no.hvl.dat251.fjelltur.dto.CreateTimeRuleRequest
 import no.hvl.dat251.fjelltur.exception.NotUniqueRuleException
-import no.hvl.dat251.fjelltur.model.DistanceRule
-import no.hvl.dat251.fjelltur.model.Rule
-import no.hvl.dat251.fjelltur.model.TimeRule
+import no.hvl.dat251.fjelltur.entity.DistanceRule
+import no.hvl.dat251.fjelltur.entity.Rule
+import no.hvl.dat251.fjelltur.entity.TimeRule
 import no.hvl.dat251.fjelltur.repository.RuleRepository
 import no.hvl.dat251.fjelltur.service.RuleService
 import org.springframework.beans.factory.annotation.Autowired
