+++ src/main/kotlin/no/hvl/dat251/fjelltur/service/impl/TripServiceImpl.kt
@@ -6,9 +6,9 @@ import no.hvl.dat251.fjelltur.exception.AccountAlreadyOnTripException
 import no.hvl.dat251.fjelltur.exception.NoCurrentTripException
 import no.hvl.dat251.fjelltur.exception.TripNotFoundException
 import no.hvl.dat251.fjelltur.exception.TripNotOngoingException
-import no.hvl.dat251.fjelltur.model.Account
-import no.hvl.dat251.fjelltur.model.GPSLocation
-import no.hvl.dat251.fjelltur.model.Trip
+import no.hvl.dat251.fjelltur.entity.Account
+import no.hvl.dat251.fjelltur.entity.GPSLocation
+import no.hvl.dat251.fjelltur.entity.Trip
 import no.hvl.dat251.fjelltur.repository.TripRepository
 import no.hvl.dat251.fjelltur.service.AccountService
 import no.hvl.dat251.fjelltur.service.TripService
